added methods in the conftest
