+++ tests/test_cases_main_page.py
@@ -1,5 +1,4 @@
 """Stores tests related to Start page and Main Page"""
-from random import random
 from time import sleep
 
 import pytest
@@ -36,38 +35,6 @@ class TestMainPage(BaseTest):
         # Logout
         main_page.logout()
         return temp_username, temp_email, temp_password
-
-    def random_username(self):
-        """Return random username"""
-        str_abc = 'qwertyuioplkjhgfdsazxcvbnm'
-        set_symbols = list(str_abc)
-        random.shuffle(set_symbols)
-        new_username = ''.join(set_symbols)
-        index = random.choice(range(3, 20))
-        return new_username.capitalize()[:index]
-
-    def random_email(self):
-        """Create random email"""
-        str_abc = 'qwertyuioplkjhgfdsazxcvbnm'
-        set_symbols = list(str_abc)
-        random.shuffle(set_symbols)
-        name = ''.join(set_symbols)
-        index = random.choice(range(3, 20))
-        email_item = ['@gmail.com', '@ukr.net', '@mail.com']
-        new_email = name[:index] + random.choice(email_item)
-        return new_email
-
-    def random_password(self):
-        """Create random password"""
-        str_abc = 'qwertyuioplkjhgfdsazxcvbnm'
-        str_abc_upper = str_abc.upper()
-        str_num = '1234567890'
-        str_symbols = '!#$%^&*()_?><'
-        str_set_symbols = str_abc + str_abc_upper + str_num + str_symbols
-        set_symbols = list(str_set_symbols)
-        random.shuffle(set_symbols)
-        new_pas = ''.join(set_symbols)
-        return new_pas[:12]
 
     def test_refresh_first_page(self, start_page, registered_user):
         """Refresh first page:
